Split pods.go into smaller parts.
